100% test coverage on employee and customer services and update those services to proper naming conventions.

+++ eagleeye-msp-quote-service/src/test/groovy/com/shew/consulting/eagleeye/msp/quote/service/model/management/pc/antivirus/detection/AntivirusDetectionSpec.groovy
@@ -16,7 +16,7 @@ class AntivirusDetectionSpec extends Specification {
         then:
         service // not null
         service.id == 'AntivirusDetectionPC'
-        service.price == 3.25
+        service.price == 3.25 as Double
         service.title == 'AntiVirus Detection'
     }
